Add new unit tests for event app

+++ vms/event/services.py
@@ -1,18 +1,30 @@
 from django.core.exceptions import ObjectDoesNotExist
 from event.models import Event
 
+#check if the event exists and is not empty
+def event_not_empty(event_id):
+    result = True
+    event = get_event_by_id(event_id)
+    if not event:
+        result = False
+    return result
+
 #need to check that this event is not accociated with any jobs, 
 #otherwise the jobs that it is associated with will be cascade deleted
 def delete_event(event_id):
 
     result = True
     event = get_event_by_id(event_id)
-    #check if there are currently any jobs associated with this event
-    jobs_in_event = event.job_set.all()
+	
+    if event_not_empty(event_id):
+        #check if there are currently any jobs associated with this event
+        jobs_in_event = event.job_set.all()
 
-    #can only delete an event if no jobs are currently associated with it
-    if event and (not jobs_in_event):
-        event.delete()
+        #can only delete an event if no jobs are currently associated with it
+        if event and (not jobs_in_event):
+            event.delete()
+        else:
+            result = False
     else:
         result = False
